move Discounts to DiscountVoucher.ts
